refactor:delete friend시nickname return
